requirements.txt
gradio>=3.50.2
pillow>=10.0.0
opencv-python==4.9.0.80
numpy==1.26.4
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0
mediapipe~=0.10.9
imageio~=2.34.0
piexif~=1.1.3